(PC-5473) Fetch the displayed modules in <Home /> instead of <HomeBody />
